chore(F-113): rename props to LayoutProps
