+++ frontend/src/app/[locale]/(main)/onboarding/page.tsx
@@ -1,6 +1,6 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
-import type { Props } from '@/interfaces/Props';
+import type { Props } from '@/interfaces/LayoutProps';
 import OnboardingPageComponent from '@/components/common/OnboardingPage';
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
