+++ frontend/src/app/[locale]/(main)/preparation/page.tsx
@@ -1,6 +1,6 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
-import type { Props } from '@/interfaces/Props';
+import type { Props } from '@/interfaces/LayoutProps';
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { locale } = await params;
