+++ frontend/src/app/[locale]/(standalone)/simulation/[id]/page.tsx
@@ -1,7 +1,7 @@
 import SimulationPageComponent from '@/components/common/SimulationPage';
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
-import type { Props } from '@/interfaces/Props';
+import type { Props } from '@/interfaces/LayoutProps';
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { locale } = await params;
