+++ frontend/src/interfaces/LayoutProps.ts
@@ -1,3 +1,4 @@
 export interface Props {
   params: Promise<{ locale: string }>;
+  children: React.ReactNode;
 }
